test pass lgtm

+++ src/main/java/com/miniproject/library/dto/loan/LoanResponse.java
@@ -15,5 +15,5 @@ public class LoanResponse {
     private Date dateReturn;
     private Date dueBorrow;
     private Integer bookCartId;
-    private Double overdueFine;
+    private Integer overdueFine;
 }
